Update client tests
